Refactor mmss app
